Split delegates into separate files for clarity
