+++ activity_browser/app/ui/tables/delegates/float.py
@@ -1,0 +1,35 @@
+# -*- coding: utf-8 -*-
+from PyQt5 import QtCore, QtGui, QtWidgets
+
+
+class FloatDelegate(QtWidgets.QStyledItemDelegate):
+    """ For managing and validating entered float values.
+    """
+    def __init__(self, parent=None):
+        super().__init__(parent)
+
+    def createEditor(self, parent, option, index):
+        editor = QtWidgets.QLineEdit(parent)
+        locale = QtCore.QLocale(QtCore.QLocale.English)
+        locale.setNumberOptions(QtCore.QLocale.RejectGroupSeparator)
+        validator = QtGui.QDoubleValidator()
+        validator.setLocale(locale)
+        editor.setValidator(validator)
+        return editor
+
+    def setEditorData(self, editor: QtWidgets.QLineEdit, index: QtCore.QModelIndex):
+        """ Populate the editor with data if editing an existing field.
+        """
+        data = index.data(QtCore.Qt.DisplayRole)
+        value = float(data) if data else 0
+        editor.setText(str(value))
+
+    def setModelData(self, editor: QtWidgets.QLineEdit, model: QtCore.QAbstractItemModel,
+                     index: QtCore.QModelIndex):
+        """ Take the editor, read the given value and set it in the model
+        """
+        try:
+            value = float(editor.text())
+            model.setData(index, value, QtCore.Qt.EditRole)
+        except ValueError:
+            pass
